% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_score}
\alias{add_score}
\title{add_score
Prend en entrée un tibble qui contient des scores et renvoie avec un score calculé dans une colonne score
Attention, la façon d'agréger les scores doit être cohérente avec la façon dont ils ont été calculés individuellement.}
\usage{
add_score(dataset, weight)
}
\arguments{
\item{dataset}{les données utilisées}

\item{weight}{un vecteur de poids constitué uniquement de 0, 1 ou -1}
}
\value{
le data_frame avec une colonne score
}
\description{
add_score
Prend en entrée un tibble qui contient des scores et renvoie avec un score calculé dans une colonne score
Attention, la façon d'agréger les scores doit être cohérente avec la façon dont ils ont été calculés individuellement.
}
