% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ansesPalette}
\alias{ansesPalette}
\title{ansesPalette}
\format{
un vecteur de nombres hexadécimaux
}
\source{
couleur prélevées par pipette sur le logo de l'anses
}
\usage{
ansesPalette
}
\description{
Palette de couleur prélevée sur le logo de l'anses
}
\keyword{datasets}
