% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{couleursCriteres}
\alias{couleursCriteres}
\title{couleursCriteres
Attribution de certaines couleurs de la palette à certains critères}
\format{
un vecteur nommé de nombres hexadécimaux
}
\usage{
couleursCriteres
}
\description{
couleursCriteres
Attribution de certaines couleurs de la palette à certains critères
}
\keyword{datasets}
