% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{criteresList}
\alias{criteresList}
\title{criteresList
Liste des critères utilisés dans la caractérisation des aliments}
\format{
un vecteur de charactères
}
\usage{
criteresList
}
\description{
criteresList
Liste des critères utilisés dans la caractérisation des aliments
}
\keyword{datasets}
