% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data_clean}
\alias{data_clean}
\title{data_clean
Les données de l'anses un peu toilettées mais non modifiées}
\format{
un tibble
}
\source{
\url{https://www.data.gouv.fr/fr/datasets/table-de-composition-nutritionnelle-des-aliments-ciqual/}
}
\usage{
data_clean
}
\description{
data_clean
Les données de l'anses un peu toilettées mais non modifiées
}
\keyword{datasets}
