% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{dataHierViz}
\alias{dataHierViz}
\title{dataHierViz
Les données de l'anses reformatées pour être utilisées dans un graphe hiérarchique type sunburst ou treemap}
\format{
un tibble avec des colonnes ids, parents et labels
}
\usage{
dataHierViz
}
\description{
dataHierViz
Les données de l'anses reformatées pour être utilisées dans un graphe hiérarchique type sunburst ou treemap
}
\keyword{datasets}
