% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{labelsHover}
\alias{labelsHover}
\title{labelsHover
Les éléments de texte qui permettront de créer les labels personnalisés}
\format{
un tibble
}
\usage{
labelsHover
}
\description{
labelsHover
Les éléments de texte qui permettront de créer les labels personnalisés
}
\keyword{datasets}
