% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{objectifCriteres}
\alias{objectifCriteres}
\title{objectifCriteres
Objectif a priori concernant chacun des critères}
\format{
un vecteur constitué de "max" ou de "min"
}
\usage{
objectifCriteres
}
\description{
objectifCriteres
Objectif a priori concernant chacun des critères
}
\keyword{datasets}
