% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{scorifyGamma}
\alias{scorifyGamma}
\title{scorifyGamma
Transforme une serie de nombre de type gamma en score
Le paramétrage est déterminé de façon à trouver un intermédiaire entre approche cardinale
et ordinale i.e. on prend un paramètre de forme >1}
\usage{
scorifyGamma(x, shape = 1)
}
\arguments{
\item{x}{un vecteur}

\item{shape}{un paramètre de forme. Egal à 1 par défaut. Comportement différent si <1 ou >1}
}
\value{
un score
}
\description{
scorifyGamma
Transforme une serie de nombre de type gamma en score
Le paramétrage est déterminé de façon à trouver un intermédiaire entre approche cardinale
et ordinale i.e. on prend un paramètre de forme >1
}
