% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{scorifyGauss}
\alias{scorifyGauss}
\title{scorifyGauss
Transforme une serie de nombre de type gaussien en score}
\usage{
scorifyGauss(x)
}
\arguments{
\item{x}{un vecteur}
}
\value{
un score
}
\description{
scorifyGauss
Transforme une serie de nombre de type gaussien en score
}
