% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{scorifyUni}
\alias{scorifyUni}
\title{scorifyUni
Transforme une serie de nombre de type uniforme en score}
\usage{
scorifyUni(x)
}
\arguments{
\item{x}{les valeurs initiales}
}
\value{
les scores
}
\description{
scorifyUni
Transforme une serie de nombre de type uniforme en score
}
