% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{scorifyUni}
\alias{scorifyUni}
\title{scorifyNonPara
Transforme une serie de nombre en score de façon non paramétrique}
\usage{
scorifyUni(x)

scorifyUni(x)
}
\arguments{
\item{x}{les valeurs initiales}
}
\value{
les scores

les scores
}
\description{
scorifyNonPara
Transforme une serie de nombre en score de façon non paramétrique

scorifyUni
Transforme une serie de nombre de type uniforme en score
}
